requirements.txt
pyusb>=1.2.1
libusb1>=3.0.0
